+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 - **errors:** Stack trace to recovered errors
 - **log:** Structured logger interface
+- **log:** Default nop logger implementation
 
 
 ## 0.9.0 - 2017-08-22