+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - **errors:** Stack trace to recovered errors
 - **log:** Structured logger interface
 - **log:** Default nop logger implementation
+- **errors:** Default nop error handler implementation
 
 
 ## 0.9.0 - 2017-08-22